redesign email html content and send data to it

+++ src/Codecool.CodecoolShop/Daos/IOrderDao.cs
@@ -1,14 +1,10 @@
 ﻿using Codecool.CodecoolShop.Models;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace Codecool.CodecoolShop.Daos
 {
     public interface IOrderDao : IDao<Item>
     {
-        int GetTotalValue();
+        decimal GetTotalValue();
         void Update(Item item, int quantity);
     }
 }
